+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/types/Multiname.java
@@ -363,7 +363,8 @@ public class Multiname {
         } else {
             String name = abc.constants.getString(name_index);
 
-            if (namespace_index > 0 && getNamespace(abc.constants).kind == Namespace.KIND_NAMESPACE) {
+            int nskind = namespace_index <= 0 ? -1 : getNamespace(abc.constants).kind;
+            if (nskind == Namespace.KIND_NAMESPACE || nskind == Namespace.KIND_PACKAGE_INTERNAL) {
                 DottedChain dc = abc.findCustomNs(namespace_index);
                 String nsname = dc != null ? dc.getLast() : null;